// frontend/src/api/model/problemDetails.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ProblemDetails {
    /** @nullable */
    type?: string | null;
    /** @nullable */
    title?: string | null;
    /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
    status?: number | string | null;
    /** @nullable */
    detail?: string | null;
    /** @nullable */
    instance?: string | null;
}